need to make more changes

+++ TransactionSim/src/main/java/com/bankSim/dto/requests/LoanPaymentRequest.java
@@ -7,14 +7,14 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 
 public class LoanPaymentRequest {
     private final Long loanId;
-    private final BigDecimal paymentAmount;
+    private final double paymentAmount;
     private final String paymentMethod;
     private final String paymentStatus;
     private final LocalDateTime paymentDate;
 
     public LoanPaymentRequest(
         @JsonProperty("loanId")Long loanId,
-        @JsonProperty("paymentAmount")BigDecimal paymentAmount,
+        @JsonProperty("paymentAmount")double paymentAmount,
         @JsonProperty("paymentMethod")String paymentMethod,
         @JsonProperty("paymentStatus")String paymentStatus
         ) {
@@ -27,7 +27,7 @@ public class LoanPaymentRequest {
 
     public Long getLoanId() {
         return loanId;
-    }public BigDecimal getPaymentAmount() {
+    }public double getPaymentAmount() {
         return paymentAmount;
     }public LocalDateTime getPaymentDate() {
         return paymentDate;
